requirements.txt
# python 3.10.8
langchain-community==0.3.15
langchain_text_splitters==0.3.5 
langchain_chroma==0.2.0 
langchain-groq==0.2.3
sentence_transformers==3.3.1 
duckduckgo-search==7.2.1 
streamlit==1.31.0
beautifulsoup4==4.12.3
python-dotenv